# api/swagger-spec.json
{"openapi":"3.0.0","paths":{"/collections/{collectionId}/items":{"post":{"operationId":"ItemsController_addItem","parameters":[{"name":"collectionId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Item"}}}},"responses":{"200":{"description":"Add new item to collection","content":{"application/json":{"schema":{"type":"number"}}}}},"tags":["Items"],"security":[{"bearer":[]}]},"get":{"operationId":"ItemsController_getItems","parameters":[{"name":"collectionId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Sorted items","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Item"}}}}}},"tags":["Items"],"security":[{"bearer":[]}]}},"/collections/{collectionId}/items/{id}/bestpair":{"get":{"operationId":"ItemsController_getBestPair","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}},{"name":"exclude","required":false,"in":"query","schema":{"type":"array","items":{"type":"number"}}}],"responses":{"200":{"description":"Get best pair for item to compare","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemPair"}}}}},"tags":["Items"],"security":[{"bearer":[]}]}},"/collections":{"get":{"operationId":"CollectionController_getCurrentUserCollections","parameters":[],"responses":{"200":{"description":"User current user collections","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Collection"}}}}}},"tags":["Collections"],"security":[{"bearer":[]}]},"post":{"operationId":"CollectionController_addCollection","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Collection"}}}},"responses":{"201":{"description":""}},"tags":["Collections"],"security":[{"bearer":[]}]}},"/collections/{collectionId}":{"get":{"operationId":"CollectionController_getCollection","parameters":[{"name":"collectionId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Get collection","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Collection"}}}}},"tags":["Collections"],"security":[{"bearer":[]}]}},"/collections/fav":{"get":{"operationId":"CollectionController_getCurrentUserFavoriteCollections","parameters":[],"responses":{"200":{"description":"User current user favorite collections","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Collection"}}}}}},"tags":["Collections"],"security":[{"bearer":[]}]}},"/relations/{fromId}/{toId}":{"post":{"operationId":"ItemRelationsController_addItem","parameters":[{"name":"fromId","required":true,"in":"path","schema":{"type":"number"}},{"name":"toId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["ItemsRelation"],"security":[{"bearer":[]}]}},"/relations/{itemAId}/{itemBId}":{"delete":{"operationId":"ItemRelationsController_getItems","parameters":[{"name":"itemAId","required":true,"in":"path","schema":{"type":"number"}},{"name":"itemBId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["ItemsRelation"],"security":[{"bearer":[]}]}}},"info":{"title":"Queuest api","description":"The queuest API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"Item":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"type":{"type":"string"}},"required":["name"]},"ItemRelation":{"type":"object","properties":{"from":{"type":"number"},"to":{"type":"number"}},"required":["from","to"]},"ItemPair":{"type":"object","properties":{"item1":{"$ref":"#/components/schemas/Item"},"item2":{"$ref":"#/components/schemas/Item"},"relation":{"$ref":"#/components/schemas/ItemRelation"}},"required":["item1","item2"]},"Collection":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"}},"required":["name"]}}}}